Added ability to create square or circle
Right now, the user can select either a square or circle,
create new square or circle via double click and selecting
one of the buttons below the canvas.

Selecting one of the shapes will highlight the shapes also.
Current issues are that the circle can only be selected and dragged
by dragging it from the bottom right quadrant of the circle. Will
be addressed later but I want to get these shapes connecting soon.