Removed coding box from index.
Currently working on the home page that will try to appease the eyes.
The coding box has been moved to the challenges -> show html.
To access it, create a challenge, and then click the link that is the
name of the challenge.

Will work on how users will access the challenges after homepage is
finished as there needs to be steps to get to such pages.