Made it to where it won't redirect to home.
It also approves if the said input is correct.

Currently working on:
Having the program find out if the person has the stuff it wants
like for loop or if there is an if-statement in there.

Brushing up the UI which shouldn't be too hard, the focus is on
making the coding part work.

Saving user's progress which I need to talk with Kyler on as
it'll be linked with scoreboard.