No need to refresh to re-render.
Will have to fast track other parts of this project as
attempting to implement a terminal has been very taxing
on time and I have almost little to none left before
we have to put it in use on Wednesday.

Essentially, the terminal/console will not be in use
or I will be implementing a very limited, basic and
rigid console for users to put in their numbers.

A real console will require a lot of time and effort to
implement, and while libraries are out there, there is
absolutely no time to study their huge documentations
to achieve what we want. And I feel rails has been
quite limiting for what we need.