Input/Output works******************
Lots of asterisks for no subtley compared to fast talking
commercials.

The Input is currently VERY limited. It will accept input
based on some arbitrary amount and will work for a
guesstimated 40% of lightweight interactive programs.
It currently only takes integers, will fix that
later. but so far, it meets what we want from
an editor that can interpret, take input and
output.

I've tested it on a few of the programs given by
professor Subbaraj and currently does the job for some.

The only issue is that that ruby's "safe" I/O that
interacts with shell and sends one output back.
So I append the entire log into one variable
and send it back for the user to see.

Users are also able to refresh console
and evaluate without the page loading. The output
box is at the bottom and the input box at the top.

Input is separated by using the comma as a delimiter.

Currently, python works to an extent so if a student
attempts to do anything with ruby, they'll be met
with disappointment and limited success.

Kyler will likely be able to do a
``{string_name}.split`` to get the answer and
add scores to the users.